Extra a helper function for tests.
This allows the tests to be more readable, by directly referring to the
expected value together with the parameters used for the test.

It uses two characteristics of C#: the fact that arrays can be defined
inline (almost as literals) and the fact that Assert.AreEqual is able to
correctly compare List and arrays based on their contents.

Having a single place where this is done because that any improvements
can be done once and it will be applied to all of its usages: for
instance I lost one property of one of your tests, which allowed the
function to return the results in either order; it can be added back by
only changing the helper function and all tests (existing or newly
added) will get the same feature.